Pest_backend/requirements.txt
fastapi
uvicorn
openai
torch
torchvision
python-multipart
python-dotenv
pillow
